Refactors authorisation methodology
