+++ src/Altinn.App.Core/Features/Correspondence/Models/CorrespondenceApiScopes.cs
@@ -1,0 +1,11 @@
+namespace Altinn.App.Core.Features.Correspondence.Models;
+
+/// <summary>
+/// Known scopes for the correspondence API.
+/// </summary>
+internal static class CorrespondenceApiScopes
+{
+    public const string ServiceOwner = "altinn:serviceowner";
+    public const string Read = "altinn:correspondence.read";
+    public const string Write = "altinn:correspondence.write";
+}
